Refactor Ctd
