+++ packages/optimizely-sdk/tests/odpEventApiManager.spec.ts
@@ -18,7 +18,7 @@
 
 import { anyString, anything, instance, mock, resetCalls, verify, when } from 'ts-mockito';
 import { LogHandler, LogLevel } from '../lib/modules/logging';
-import { OdpEventApiManager } from '../lib/core/odp/event_manager/odp_event_api_manager';
+import { OdpEventApiManager } from '../lib/core/odp/odp_event_api_manager';
 import { OdpEvent } from '../lib/core/odp/odp_event';
 import { RequestHandler } from '../lib/utils/http_request_handler/http';
